requirements.txt
discord.py
tradingview_ta
python-dotenv
